Moved graphql queries to use hasura only removed auth server (#277)
* Moved graphql queries to use hasura only removed auth server

* removed process.env.REACT_APP_AUTH_SERVER_GRAPHQL_URL
