+++ front-end/codegen.yml
@@ -5,23 +5,9 @@ generates:
       - ${REACT_APP_HASURA_GRAPHQL_URL}:
           headers:
             X-Hasura-Admin-Secret: ${HASURA_GRAPHQL_ADMIN_SECRET}
-    documents: 
+    documents:
       - "src/**/*.{ts,tsx}"
       - "!src/services/*"
-    plugins:
-      - "typescript"
-      - "typescript-operations"
-      - "typescript-react-apollo"
-    config:
-      withComponent: false
-      withHOC: false
-      withHooks: true
-      withResultType: true
-  src/generated/auth-graphql.tsx:
-    schema:
-      - ${REACT_APP_AUTH_SERVER_GRAPHQL_URL}
-    documents: 
-      - "src/services/**/*.{ts,tsx}"
     plugins:
       - "typescript"
       - "typescript-operations"
