+++ front-end/src/components/queries.ts
@@ -1,0 +1,9 @@
+import gql from 'graphql-tag';
+
+export const GET_REFRESH_TOKEN = gql`
+query GET_REFRESH_TOKEN {
+    token {
+        token
+    }
+}
+`;
