+++ front-end/src/screens/LoginForm/query.ts
@@ -1,0 +1,16 @@
+import gql from 'graphql-tag';
+
+export const LOGIN = gql`
+    mutation LOGIN($password: String!, $username: String!) {
+        login(password: $password, username:$username){
+            user {
+                id
+                name
+                username
+                email
+                email_verified
+            }
+            token
+        }
+    }
+`;
