+++ front-end/src/screens/MenuBar/query.ts
@@ -1,0 +1,9 @@
+import gql from 'graphql-tag';
+
+export const LOGOUT = gql`
+    mutation LOGOUT {
+        logout {
+            message
+        }
+    }
+`;
